change blogpost

+++ blog_posts/views.py
@@ -1,37 +1,34 @@
 from django.contrib import messages
 from django.shortcuts import render, get_object_or_404, redirect
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponseRedirect, Http404
+from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
+from django.utils import timezone
+from django.db.models import Q
 from .models import BlogPost
 from .forms import BlogPostForm
 
 
 def post_create(request):
-    form = BlogPostForm(request.POST or None)
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    form = BlogPostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
+        instance.user = request.user
         instance.save()
         messages.success(request, 'Пост создан!')
         return HttpResponseRedirect(instance.get_absolute_url())
-    else:
-        messages.error(request, 'Пост не был создан!')
     context = {
         'form': form,
     }
     return render(request, 'post_form.html', context)
 
 
-def post_list(request):
-    queryset = BlogPost.objects.all()
-    context = {
-        'object_list': queryset,
-        'title': 'List',
-    }
-    return render(request, 'blog.html', context)
-
-
 def post_update(request, id=id):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
     instance = get_object_or_404(BlogPost, id=id)
-    form = BlogPostForm(request.POST or None, instance=instance)
+    form = BlogPostForm(request.POST or None, request.FILES or None, instance=instance)
     if form.is_valid():
         instance = form.save(commit=False)
         instance.save()
@@ -44,18 +41,58 @@ def post_update(request, id=id):
     return render(request, 'post_form.html', context)
 
 
+def post_delete(request, id=id):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    instance = get_object_or_404(BlogPost, id=id)
+    instance.delete()
+    messages.success(request, 'Пост удалён!')
+    return redirect('blog:list')
+
+
 def post_detail(request, id=id):
     instance = get_object_or_404(BlogPost, id=id)
+    today = timezone.now().date()
+    if instance.draft or instance.publish > timezone.now().date():
+        if not request.user.is_staff or not request.user.is_superuser:
+            raise Http404
     context = {
         'title': instance.title,
         'instance': instance,
+        'today': today,
     }
     return render(request, 'post_detail.html', context)
 
 
-def post_delete(request, id=id):
-    instance = get_object_or_404(BlogPost, id=id)
-    instance.delete()
-    messages.success(request, 'Пост создан!')
-    return redirect('blog:list')
+def post_list(request):
+    queryset_list = BlogPost.objects.active()
+    today = timezone.now().date()
+    if request.user.is_staff or request.user.is_superuser:
+        queryset_list = BlogPost.objects.all()
+    query = request.GET.get('q')
+    if query:
+        queryset_list = queryset_list.filter(
+            Q(title__icontains=query)|
+            Q(content__icontains=query)|
+            Q(user__first_name__icontains=query)|
+            Q(user__last_name__icontains=query)
+        ).distinct()
+    paginator = Paginator(queryset_list, 5)
+    page_request_var = 'page'
+    page = request.GET.get(page_request_var)
+    try:
+        queryset = paginator.page(page)
+    except PageNotAnInteger:
+        # If page is not an integer, deliver first page.
+        queryset = paginator.page(1)
+    except EmptyPage:
+        # If page is out of range (e.g. 9999), deliver last page of results.
+        queryset = paginator.page(paginator.num_pages)
+    context = {
+        'object_list': queryset,
+        'title': 'List',
+        'page_request_var':  page_request_var,
+        'today': today,
+    }
+    return render(request, 'blog.html', context)
